feat(remote-context): add _address_of timeout
Prior to this commit, calls to `_address_of` with an unrecognized label
resulted in an infinite blocking loop waiting for a response.  Since the
label was unrecognized, no response is ever sent from the device.

As of this commit, an optional timeout may be specified for the
`_address_of` method.  If no response is received before the specified
timeout period, a `RuntimeError` exception is raised.